add edit form for relations
